formatted, refactored, and divided the ORM classes up into files for Single Responsibility reason
